Minor changes to PPSD and save_npz_min
Added the winparams option to PPSD: if equal to 'obspy-cosine' (or not specified) uses the original 10% cosine taper, otherwise the parameters to be specified are the ones used by scipy.signal.get_window.
Added path and fmt options to save_npz_min to specify the folder where to save the results and the format of the name: "comp" uses the year.julday, "ext" uses year.month.day. The network is omitted to have a naming system coherent with the waveform.